Add XR_HTC_passthrough
The problem with XR_HTC_passthrough was that we were generating code
for XrCompositionLayerPassthroughHTC that didn't compile.
This erroneous code was in the high level bindings so it turn out that
we don't need to disable the whole extension if we just skip the high
level bindings for XrCompositionLayerPassthroughHTC.

This also removes the ugly blacklist for structs during parsing that
was added to avoid the problematic code gen.