Don't fail if `xrEnumerateApiLayerProperties` is not available
Honestly, this is not very nice, but the fact is both monado and SteamVR fails
if their `xrGetInstanceProcAddr` functions are used to query
"xrEnumerateApiLayerProperties". Monado returns `ERROR_NOT_SUPPORTED` if
instance is not NULL, but will return a valid function if it is. On the
other hand SteamVR returns `ERROR_INVALID_HANDLE` if instance _is_ NULL.

Given it doesn't really make sense to query API layers from an API
layer, allow queries of "xrEnumerateApiLayerProperties" to fail.